Support the refresh_token grant when user specifies `--renew`.
Also remove the SFCC_CLIENT_RENEW_TOKEN boolean config parameter, because we can compute this on the fly by not storing the refresh token if refresh isn't desired.

Also support detection of expired token via `body.error == invalid_token`.
The CCDX API server sends a `error` response in the body when the token is expired, in contrast to AM sending a `fault` body. We now support both.