fixes #280 node-fetch > 3 requires ESM; 2.9.0 broken
node-fetch > 3 requires ESM but was updated by snyk automatically; this
breaks anything using the lib/slas.js file (as well as projects using
this as a library). i.e. 2.9.0 is broken in many use cases due to this
dependency.

2.6.7 is fixed anyway according to the snyk vulnerability that triggered
the initial upgrade of the major version
(https://security.snyk.io/vuln/SNYK-JS-NODEFETCH-2342118) so this is
unlikely to reoccur. However we may also want to prevent it from
trying a major update again via configuration.